refactor: Timer 하나의 디렉토리롤 빼서 관리
